Split header files into .h and .cpp
