+++ includes/FileSystem.h
@@ -3,7 +3,6 @@
 
 #include <string>
 #include <cstdlib>
-#include "root_directory.h"
 
 class FileSystem
 {
@@ -11,42 +10,13 @@ private:
     typedef std::string (*Builder)(const std::string &path);
 
 public:
-    static std::string getPath(const std::string &path)
-    {
-        static std::string (*pathBuilder)(std::string const &) = getPathBuilder();
-        return (*pathBuilder)(path);
-    }
+    static std::string getPath(const std::string &path);
 
 private:
-    static std::string const &getRoot()
-    {
-        static char const *envRoot = getenv("LOGL_ROOT_PATH");
-        static char const *givenRoot = (envRoot != nullptr ? envRoot : logl_root);
-        static std::string root = (givenRoot != nullptr ? givenRoot : "");
-        return root;
-    }
-
-    static Builder getPathBuilder()
-    {
-        if (getRoot() != "")
-        {
-            return &FileSystem::getPathRelativeRoot;
-        }
-        else
-        {
-            return &FileSystem::getPathRelativeBinary;
-        }
-    }
-
-    static std::string getPathRelativeRoot(const std::string &path)
-    {
-        return (getRoot() + std::string("/") + path);
-    }
-
-    static std::string getPathRelativeBinary(const std::string &path)
-    {
-        return ("../../../" + path);
-    }
+    static std::string const &getRoot();
+    static Builder getPathBuilder();
+    static std::string getPathRelativeRoot(const std::string &path);
+    static std::string getPathRelativeBinary(const std::string &path);
 };
 
 #endif
