+++ includes/Shader.h
@@ -13,150 +13,27 @@
 class Shader
 {
 public:
-	unsigned int ID; // shader program ID
-
-	Shader(const char *vertexPath, const char *fragmentPath) // creates and links shader program
-	{
-		// retrieve from file path
-		std::string vertexCode;
-		std::string fragmentCode;
-		std::ifstream vShaderFile;
-		std::ifstream fShaderFile;
-
-		// ensure exceptions
-		vShaderFile.exceptions(std::ifstream::failbit | std::ifstream::badbit);
-		fShaderFile.exceptions(std::ifstream::failbit | std::ifstream::badbit);
-		try
-		{
-			// open files
-			vShaderFile.open(vertexPath);
-			fShaderFile.open(fragmentPath);
-			std::stringstream vShaderStream, fShaderStream;
-			// read buffers into stream
-			vShaderStream << vShaderFile.rdbuf();
-			fShaderStream << fShaderFile.rdbuf();
-			//close file handlers
-			vShaderFile.close();
-			fShaderFile.close();
-			// convert stream to string
-			vertexCode = vShaderStream.str();
-			fragmentCode = fShaderStream.str();
-		}
-		catch (std::ifstream::failure e)
-		{
-			std::cout << "ERROR::SHADER::FILE_NOT_SUCEESFULLY_READ" << std::endl;
-		}
-
-		const char *vShaderCode = vertexCode.c_str();
-		const char *fShaderCode = fragmentCode.c_str();
-		unsigned int vertex, fragment;
-		// vertex Shader
-		vertex = glCreateShader(GL_VERTEX_SHADER);
-		glShaderSource(vertex, 1, &vShaderCode, NULL);
-		glCompileShader(vertex);
-		checkCompileErrors(vertex, "VERTEX");
-		// fragment Shader
-		fragment = glCreateShader(GL_FRAGMENT_SHADER);
-		glShaderSource(fragment, 1, &fShaderCode, NULL);
-		glCompileShader(fragment);
-		checkCompileErrors(fragment, "FRAGMENT");
-
-		// shader program
-		ID = glCreateProgram();
-		glAttachShader(ID, vertex);
-		glAttachShader(ID, fragment);
-		glLinkProgram(ID);
-		checkCompileErrors(ID, "PROGRAM");
-		// delete individual shaders
-		glDeleteShader(vertex);
-		glDeleteShader(fragment);
-	}
-
+	unsigned int ID;										  // shader program ID
+	Shader(const char *vertexPath, const char *fragmentPath); // creates and links shader program
 	// uses shader program
-	void use()
-	{
-		glUseProgram(ID);
-	}
-
+	void use();
 	// utility functions
-	void setBool(const std::string &name, bool value) const
-	{
-		glUniform1i(glGetUniformLocation(ID, name.c_str()), (int)value);
-	}
-	void setInt(const std::string &name, int value) const
-	{
-		glUniform1i(glGetUniformLocation(ID, name.c_str()), value);
-	}
-	void setFloat(const std::string &name, float value) const
-	{
-		glUniform1f(glGetUniformLocation(ID, name.c_str()), value);
-	}
-	void setVec2(const std::string &name, const glm::vec2 &value) const
-	{
-		glUniform2fv(glGetUniformLocation(ID, name.c_str()), 1, &value[0]);
-	}
-	void setVec2(const std::string &name, float x, float y) const
-	{
-		glUniform2f(glGetUniformLocation(ID, name.c_str()), x, y);
-	}
-	void setVec3(const std::string &name, const glm::vec3 &value) const
-	{
-		glUniform3fv(glGetUniformLocation(ID, name.c_str()), 1, &value[0]);
-	}
-	void setVec3(const std::string &name, float x, float y, float z) const
-	{
-		glUniform3f(glGetUniformLocation(ID, name.c_str()), x, y, z);
-	}
-	void setVec4(const std::string &name, const glm::vec4 &value) const
-	{
-		glUniform4fv(glGetUniformLocation(ID, name.c_str()), 1, &value[0]);
-	}
-	void setVec4(const std::string &name, float x, float y, float z, float w) const
-	{
-		glUniform4f(glGetUniformLocation(ID, name.c_str()), x, y, z, w);
-	}
-	void setMat2(const std::string &name, const glm::mat2 &mat) const
-	{
-		glUniformMatrix2fv(glGetUniformLocation(ID, name.c_str()), 1, GL_FALSE, &mat[0][0]);
-	}
-	void setMat3(const std::string &name, const glm::mat3 &mat) const
-	{
-		glUniformMatrix3fv(glGetUniformLocation(ID, name.c_str()), 1, GL_FALSE, &mat[0][0]);
-	}
-	void setMat4(const std::string &name, const glm::mat4 &mat) const
-	{
-		glUniformMatrix4fv(glGetUniformLocation(ID, name.c_str()), 1, GL_FALSE, &mat[0][0]);
-	}
+	void setBool(const std::string &name, bool value) const;
+	void setInt(const std::string &name, int value) const;
+	void setFloat(const std::string &name, float value) const;
+	void setVec2(const std::string &name, const glm::vec2 &value) const;
+	void setVec2(const std::string &name, float x, float y) const;
+	void setVec3(const std::string &name, const glm::vec3 &value) const;
+	void setVec3(const std::string &name, float x, float y, float z) const;
+	void setVec4(const std::string &name, const glm::vec4 &value) const;
+	void setVec4(const std::string &name, float x, float y, float z, float w) const;
+	void setMat2(const std::string &name, const glm::mat2 &mat) const;
+	void setMat3(const std::string &name, const glm::mat3 &mat) const;
+	void setMat4(const std::string &name, const glm::mat4 &mat) const;
 
 private:
 	// utility function for checking errors
-	void checkCompileErrors(unsigned int shader, std::string type)
-	{
-		int success;
-		char infoLog[1024];
-		if (type != "PROGRAM")
-		{
-			glGetShaderiv(shader, GL_COMPILE_STATUS, &success);
-			if (!success)
-			{
-				glGetShaderInfoLog(shader, 1024, NULL, infoLog);
-				std::cout << "ERROR::SHADER::VERTEX::COMPILATION_FAILED of type " << type << "\n"
-						  << infoLog << "\n"
-						  << std::endl;
-			}
-		}
-		else
-		{
-			glGetProgramiv(shader, GL_LINK_STATUS, &success);
-			if (!success)
-			{
-				glGetProgramInfoLog(shader, 1024, NULL, infoLog);
-				std::cout << "ERROR::SHADER::PROGRAM::LINKING_FAILED of type " << type << "\n"
-						  << infoLog << "\n"
-						  << std::endl;
-			}
-		}
-	}
+	void checkCompileErrors(unsigned int shader, std::string type);
 };
 
 #endif // !SHADER_H
