+++ includes/Utils.h
@@ -8,50 +8,7 @@
 #include <iostream>
 #include <string>
 
-unsigned int LoadTexture(const char *path)
-{
-    unsigned int textureID;
-    glGenTextures(1, &textureID);
+unsigned int LoadTexture(const char *path);
+void Log(const char *message);
 
-    int width, height, nrComponents;
-    unsigned char *data = stbi_load(path, &width, &height, &nrComponents, 0);
-
-    // setup data
-    if (data)
-    {
-        GLenum format;
-        if (nrComponents == 1)
-        {
-            format = GL_RED;
-        }
-        else if (nrComponents == 3)
-        {
-            format = GL_RGB;
-        }
-        else if (nrComponents == 4)
-        {
-            format = GL_RGBA;
-        }
-        glBindTexture(GL_TEXTURE_2D, textureID);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB, width, height, 0, format, GL_UNSIGNED_BYTE, data);
-        glGenerateMipmap(GL_TEXTURE_2D);
-
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_REPEAT);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_REPEAT);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR_MIPMAP_LINEAR);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-    }
-    else
-    {
-        std::cout << "Failed to load texture" << std::endl;
-    }
-    stbi_image_free(data);
-
-    return textureID;
-}
-
-void Log(const char *message)
-{
-    std::cout << message << "\n";
-}
 #endif
